refactor(gateway-tests): extract shared HTTP test helpers, fix review issues
http-test-helpers.ts: makeReq/makeRes/makeContext を共通化。
api-cron.test.ts: PUT で reloadScheduler 確認追加、POST でデフォルト値確認
追加（untitled/enabled:true）、不要な method 代入を削除。
api-org.test.ts: 共通ヘルパーを使うよう更新。

/review フィードバック反映
